Refactor the edge diversity/pruning computation to make the code clearer
